Consolidate RayCaster sensor tests
